feat: add current player index

+++ lib/battle/domain/entities/monster_entity.dart
@@ -6,7 +6,7 @@ part 'monster_entity.freezed.dart';
 class MonsterEntity with _$MonsterEntity {
   const factory MonsterEntity({
     required int speed,
-    required int healthPoints,
+    required double healthPoints,
     required int damage,
     required String background,
     required String attackAnimation,
