+++ lib/battle/domain/entities/players_entity.dart
@@ -5,7 +5,7 @@ part 'players_entity.freezed.dart';
 @freezed
 class PlayersEntity with _$PlayersEntity {
   const factory PlayersEntity({
-    required int healthPoints,
+    required double healthPoints,
     required int numberOfPlayers,
     required int damage,
   }) = _PlayersEntity;
